backend/requirements.txt
flask
flask-cors
transformers
torch
accelerate
bitsandbytes 
python-dotenv 